+++ src/main/java/duke/TaskList.java
@@ -2,7 +2,7 @@ package duke;
 
 import java.util.ArrayList;
 
-import task.Task;
+import duke.task.Task;
 
 public class TaskList extends ArrayList<Task>{
