+++ src/main/java/duke/Ui.java
@@ -2,7 +2,7 @@ package duke;
 
 import java.util.Scanner;
 
-import task.Task;
+import duke.task.Task;
 
 public class Ui{
     private Scanner sc;
