+++ src/main/java/duke/command/Command.java
@@ -1,4 +1,4 @@
-package command;
+package duke.command;
 
 import duke.DukeException;
 import duke.Storage;
