+++ src/main/java/duke/command/Exit.java
@@ -1,4 +1,4 @@
-package command;
+package duke.command;
 
 import duke.DukeException;
 import duke.Storage;
